Catalogue of statistics related to a variable

Retrieves list of statistics based on variable name

catalogue_statistics_by_variable(
  name,
  selection = NULL,
  area = c("free", "user", "all"),
  searchcriterion = c("code", "content"),
  sortcriterion = c("code", "content"),
  pagelength = 100,
  language = getOption("genesis_language"),
  genesis = getOption("genesis")
)

Arguments

name

Code of the variable (e.g. retrieved using catalogue_variables)

selection

String to filter for the code or content of the object(s) to be returned. Use of wildcard (*) possible.

area

Area in which the object is stored:

"free" = objects in the public catalogue

"user" = objects saved to "Meine Tabellen"

searchcriterion

Does selection refer to the code or the content of the object(s)?

sortcriterion

Should the results be sorted by the code or the content of the object(s)?

pagelength

Maximum number of results delivered

language

Search terms, returned messages and data descriptions in German ("de") or English ("en")?

A default value can also be set to the global option genesis_language via options(), e.g. options(genesis_language = "de").

genesis

Which GENESIS database should be used?

One of: "destatis", "regionalstatistik", "bildungsmonitoring", "bayern", "nrw", "sachsen-anhalt".

A default value can also be set to the global option genesis via options(), e.g. options(genesis = "destatis"), so that genesis does not have to be specified in the function call.

Value

A data.frame (or tbl_df if tibble package is installed)

Examples

if (FALSE) {
options(genesis = "destatis")
catalogue_statistics_by_variable("KREISE")
}