Table catalogue

Retrieves list of tables according to selection

catalogue_tables(
  selection,
  area = c("free", "user", "all"),
  pagelength = 100,
  language = "en",
  genesis = getOption("genesis")
)

Arguments

selection
String to filter for the code of the table(s) to be returned. Use of wildcard (*) possible.

area
Area in which the object is stored:

"free" = objects in the public catalogue

"user" = objects saved to "Meine Tabellen"

pagelength
Maximum number of results delivered

language
Search terms, returned messages and data descriptions in German ("de") or English ("en")?

genesis
Which GENESIS database should be used?

One of: "destatis", "regionalstatistik", "bildungsmonitoring", "bayern", "nrw", "sachsen-anhalt".

A default value can also be set to the global option genesis via options(), e.g. options(genesis = "destatis"), so that genesis does not have to be specified in the function call.

Value

A data.frame (or tbl_df if tibble package is installed)

Examples

if (FALSE) {
options(genesis = "destatis")
catalogue_tables("124*")
}