gen_auth_save

Save credentials of the different databases for further convenient use

Usage

gen_auth_save(database = c("all", "genesis", "zensus", "regio"))

Arguments

database

Character string. The database to store credentials for ('all', 'genesis', 'zensus' or 'regio').

Value

Path to the RDS file in which credentials are saved, invisibly.

Details

Username and password are encrypted and saved as RDS in the package config directory. A random string is generated and stored in the session environment variable GENESIS_KEY. This string is used as the key to encrypt and decrypt the entered credentials. To avoid havding to save authentication in future sessions, GENESIS_KEY can be added to .Renviron. The usethis package includes a helper function for editing .Renviron files from an R session with usethis::edit_r_environ().

Examples

if (FALSE) { # \dontrun{
gen_auth_save("zensus")
} # }