gen_download_job

gen_download_job

Usage

gen_download_job(
  name,
  database = c("genesis", "regio"),
  area = c("all", "public", "user"),
  compress = FALSE,
  language = Sys.getenv("GENESIS_LANG"),
  all_character = TRUE
)

Arguments

name

Character string. The job code retrieved by using gen_list_jobs().

database

Character string. Indicator if the GENESIS ('genesis') or regionalstatistik.de ('regio') database is called. Only one database can be addressed per function call. Default option is 'genesis'.

area

Character string. Indicator from which area of the database the results are called. In general, 'all' is the appropriate solution. Default option is 'all'.

compress

Boolean. Should empty rows and columns be discarded? Default is FALSE.

language

Character string. Defines if the decimal mark and grouping mark of integers should be represented based on the European (e.g.: '100,5', '200.000,5') or American ('100.5', '200,000.5') system. Defaults to 'Sys.getenv("GENESIS_LANG")'.

all_character

Boolean. Should all variables be imported as 'character' variables? Avoids fuzzy data type conversions if there are leading zeros or other special characters. Defaults to TRUE.

Value

Returns a data.frame with the table content

Examples

if (FALSE) { # \dontrun{
gen_download_job("21311-00-01-1_123456789", "regio")
} # }