Search for Meta-Information for a Cube

Function to search for meta-information for a specific cube.

Usage

gen_metadata_cube(code = NULL, error.ignore = FALSE, ...)

Arguments

code

a string with a maximum length of 15 characters. Code from a GENESIS-Object. Only one code per iteration.

error.ignore

a logical. Indicator if the function should stop if an error occurs or no object for the request is found or if it should produce a token as response.

...

Additional parameter of the GENESIS API call. These parameters are only affecting the GENESIS API call itself, no further processing.

Value

A list with all recalled elements from GENESIS. Attributes are added to the dataframe describing the search configuration for the returned output.

Examples

if (FALSE) {
# Find meta-information of the cube with the code "11111KE001"
object <- gen_metadata_cube(code = "11111KE001")
}