restatis: R Wrapper to Access a Wide Range of Germany's Federal Statistical System Databases Based on the GENESIS Web Service RESTful API of the German Federal Statistical Office (Statistisches Bundesamt/Destatis)

A RESTful API wrapper for accessing the GENESIS database of the German Federal Statistical Office (Destatis) as well as its Census Database and the database of Germany's regional statistics. Supports data search functions, credential management, result caching, and handling remote background jobs for large datasets.

See also

Author

Maintainer: Yannik Buhl ybuhl@posteo.de

Authors:

Zoran Kovacevic (ORCID)

Dorian Le Jeune

Long Nguyen long.nguyen@uni-bielefeld.de (ORCID)

Johannes Ritter ritter.johannes@gmail.com